search dates

+++ src/main/java/com/avlija/hotel/repository/ReservationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.avlija.hotel.model.Reservation;
 
-public interface ReservationRepository extends CrudRepository<Reservation, Integer> {
+public interface ReservationRepository extends CrudRepository<Reservation, Long> {
 
 }
